# openapi.yaml
openapi: 3.0.0
info:
  title: API de Charadas
  version: 1.0.0
  description: API para gerenciamento de charadas com autenticação via token

servers:
  - url: http://localhost:5000

paths:
  /:
    get:
      summary: Rota inicial
      responses:
        '200':
          description: Informações da API
          content:
            application/json:
              schema:
                type: object
                properties:
                  api:
                    type: string
                  version:
                    type: string
                  Author:
                    type: string

  /login:
    post:
      summary: Login do administrador
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - usuario
                - senha
              properties:
                usuario:
                  type: string
                senha:
                  type: string
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
        '400':
          description: Erro na requisição

  /charadas:
    get:
      summary: Listar todas as charadas
      responses:
        '200':
          description: Lista de charadas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Charada'

    post:
      summary: Criar nova charada
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - pergunta
                - resposta
              properties:
                pergunta:
                  type: string
                resposta:
                  type: string
      responses:
        '201':
          description: Charada criada
        '400':
          description: Erro na criação

  /charadas/aleatorias:
    get:
      summary: Retornar uma charada aleatória
      responses:
        '200':
          description: Charada aleatória
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charada'

  /charadas/{id}:
    get:
      summary: Buscar charada por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Charada encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charada'
        '404':
          description: Charada não encontrada

    put:
      summary: Atualizar completamente uma charada
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - pergunta
                - resposta
              properties:
                pergunta:
                  type: string
                resposta:
                  type: string
      responses:
        '200':
          description: Atualizado com sucesso
        '400':
          description: Erro na atualização

    patch:
      summary: Atualizar parcialmente uma charada
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pergunta:
                  type: string
                resposta:
                  type: string
      responses:
        '200':
          description: Atualizado com sucesso
        '400':
          description: Erro na atualização

    delete:
      summary: Deletar uma charada
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deletado com sucesso
        '404':
          description: Charada não encontrada

components:
  schemas:
    Charada:
      type: object
      properties:
        id:
          type: integer
        pergunta:
          type: string
        resposta:
          type: string

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT